# docs/api/mobile-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rent A 4Wheel native API",
    "version": "1.0.0",
    "description": "Phase 7A staging foundation. Live finance is disabled. No administrative mutation surface."
  },
  "servers": [
    {
      "url": "https://staging.example.invalid",
      "description": "Replace with the approved staging origin; no production approval implied"
    }
  ],
  "components": {
    "responses": {
      "TypedError": {
        "description": "Typed error; no private diagnostic payload",
        "headers": {
          "X-API-Version": {
            "schema": {
              "type": "string",
              "const": "1"
            },
            "description": "Contract version"
          },
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Server-generated correlation ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "private, no-store"
            },
            "description": "Private responses must not be cached"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "data",
                "error",
                "requestId"
              ],
              "properties": {
                "data": {
                  "type": "null"
                },
                "error": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "INVALID_REQUEST",
                        "NOT_FOUND",
                        "CONFLICT",
                        "RATE_LIMITED",
                        "UNAVAILABLE"
                      ]
                    }
                  }
                },
                "requestId": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "nativeBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque ma_ credential",
        "description": "Five-minute database-authorized device credential"
      }
    }
  },
  "paths": {
...
    "/api/v1/mobile/reservations/{id}": {
      "get": {
        "operationId": "reservation",
        "security": [
          {
            "nativeBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "const": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Completed request; authority and availability remain subject to current state",
            "headers": {
              "X-API-Version": {
                "schema": {
                  "type": "string",
                  "const": "1"
                },
                "description": "Contract version"
              },
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Server-generated correlation ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                },
                "description": "Private responses must not be cached"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "data",
                    "error",
                    "requestId"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128
                        },
                        "confirmationNumber": {
                          "type": "string"
                        },
                        "vehicleId": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128
                        },
                        "status": {
                          "type": "string"
                        },
                        "pickupAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                        },
                        "returnAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                        },
                        "bookingTimezone": {
                          "type": "string"
                        },
                        "expiresAt": {
                          "anyOf": [
                            {
                              "type": "string",
                              "format": "date-time",
                              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "subtotalCents": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "extrasCents": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "discountCents": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "taxCents": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "feesCents": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "totalCents": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "depositCents": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "bookingFingerprint": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "confirmationNumber",
                        "vehicleId",
                        "status",
                        "pickupAt",
                        "returnAt",
                        "bookingTimezone",
                        "expiresAt",
                        "subtotalCents",
                        "extrasCents",
                        "discountCents",
                        "taxCents",
                        "feesCents",
                        "totalCents",
                        "depositCents",
                        "bookingFingerprint"
                      ],
                      "additionalProperties": false
                    },
                    "error": {
                      "type": "null"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/TypedError"
          },
          "401": {
            "$ref": "#/components/responses/TypedError"
          },
          "403": {
            "$ref": "#/components/responses/TypedError"
          },
          "404": {
            "$ref": "#/components/responses/TypedError"
          },
          "409": {
            "$ref": "#/components/responses/TypedError"
          },
          "413": {
            "$ref": "#/components/responses/TypedError"
          },
          "415": {
            "$ref": "#/components/responses/TypedError"
          },
          "429": {
            "$ref": "#/components/responses/TypedError"
          },
          "500": {
            "$ref": "#/components/responses/TypedError"
          },
          "503": {
            "$ref": "#/components/responses/TypedError"
          }
        }
      }
    },
...
    "/api/v1/mobile/documents": {
      "get": {
        "operationId": "documents",
        "security": [
          {
            "nativeBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "const": "1"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Completed request; authority and availability remain subject to current state",
            "headers": {
              "X-API-Version": {
                "schema": {
                  "type": "string",
                  "const": "1"
                },
                "description": "Contract version"
              },
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Server-generated correlation ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                },
                "description": "Private responses must not be cached"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "data",
                    "error",
                    "requestId"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 128
                              },
                              "type": {
                                "type": "string"
                              },
                              "status": {
                                "type": "string"
                              },
                              "malwareScanStatus": {
                                "type": "string"
                              },
                              "reservationId": {
                                "anyOf": [
                                  {
                                    "type": "string",
                                    "minLength": 1,
                                    "maxLength": 128
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              }
                            },
                            "required": [
                              "id",
                              "type",
                              "status",
                              "malwareScanStatus",
                              "reservationId"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "nextCursor": {
                          "anyOf": [
                            {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 128
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      },
                      "required": [
                        "items",
                        "nextCursor"
                      ],
                      "additionalProperties": false
                    },
                    "error": {
                      "type": "null"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/TypedError"
          },
          "401": {
            "$ref": "#/components/responses/TypedError"
          },
          "403": {
            "$ref": "#/components/responses/TypedError"
          },
          "404": {
            "$ref": "#/components/responses/TypedError"
          },
          "409": {
            "$ref": "#/components/responses/TypedError"
          },
          "413": {
            "$ref": "#/components/responses/TypedError"
          },
          "415": {
            "$ref": "#/components/responses/TypedError"
          },
          "429": {
            "$ref": "#/components/responses/TypedError"
          },
          "500": {
            "$ref": "#/components/responses/TypedError"
          },
          "503": {
            "$ref": "#/components/responses/TypedError"
          }
        }
      }
    },
...
  }
}